git clone: move `absolute_git_source` from cli to lib and rename to `sanitize_git_url_if_path`
